change intger field to float field

+++ trips/forms.py
@@ -7,8 +7,8 @@ class SearchForm(forms.Form):
     Date2 = forms.DateField(required=False, widget=forms.DateInput(attrs={'class': 'departure', 'placeholder': 'To..'}), label='')
     duration_in_days = forms.CharField(required=False, widget=forms.NumberInput(attrs={'placeholder': 'From...'}))
     duration_in_days2 = forms.CharField(required=False, widget=forms.NumberInput(attrs={'placeholder': 'To..'}), label='')
-    price = forms.IntegerField(required=False, widget=forms.NumberInput(attrs={'placeholder': 'From...'}))
-    price2 = forms.IntegerField(required=False, widget=forms.NumberInput(attrs={'placeholder': 'To..'}), label='')
+    price = forms.FloatField(required=False, widget=forms.NumberInput(attrs={'placeholder': 'From...'}))
+    price2 = forms.FloatField(required=False, widget=forms.NumberInput(attrs={'placeholder': 'To..'}), label='')
     meals = forms.ChoiceField(required=False, choices=[('meals...', 'meals...'),('halfboard', 'halfboard'), ('fullboard', 'fullboard')], widget=forms.Select(attrs={'class': 'half'}))
     agency = forms.CharField(required=False, widget=forms.TextInput(attrs={'class': 'half', 'placeholder': 'Agnecy Name...'}))
